# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_importFile": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_importFile",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportResponse": {
        "properties": {
          "deduplicated": {
            "description": "True when the file was already on Babel",
            "title": "Deduplicated",
            "type": "boolean"
          },
          "item": {
            "$ref": "#/components/schemas/LibraryItemResponse"
          }
        },
        "required": [
          "item",
          "deduplicated"
        ],
        "title": "ImportResponse",
        "type": "object"
      },
...
      "LinkRequest": {
        "properties": {
          "url": {
            "maxLength": 2000,
            "minLength": 8,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "LinkRequest",
        "type": "object"
      },
...
      "WithdrawRequest": {
        "properties": {
          "block": {
            "default": true,
            "title": "Block",
            "type": "boolean"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "WithdrawRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Library, annotated reading, sync and statistics.",
    "title": "Babel API",
    "version": "0.25.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/admin/files/{sha256}/withdraw": {
      "post": {
        "description": "Withdraw a file from every library and delete it; by default its hash is blocked.",
        "operationId": "withdrawFile",
        "parameters": [
          {
            "in": "path",
            "name": "sha256",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Sha256",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Withdraw File",
        "tags": [
          "library",
          "admin"
        ]
      }
    },
...
    "/v1/files/{sha256}": {
      "get": {
        "description": "Download a stored file. Supports HTTP range requests to resume downloads.",
        "operationId": "downloadFile",
        "parameters": [
          {
            "in": "path",
            "name": "sha256",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Sha256",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download File",
        "tags": [
          "library"
        ]
      }
    },
    "/v1/files/{sha256}/cbz": {
      "get": {
        "description": "A comic as CBZ: CBR (RAR) files are converted once, for readers that only open ZIP.",
        "operationId": "downloadFileAsCbz",
        "parameters": [
          {
            "in": "path",
            "name": "sha256",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Sha256",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.comicbook+zip": {}
            },
            "description": "Successful Response"
          },
          "415": {
            "description": "Not a comic, or this one cannot be converted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download As Cbz",
        "tags": [
          "library"
        ]
      }
    },
    "/v1/files/{sha256}/cover": {
      "get": {
        "description": "The cover found in a stored file (EPUB, CBZ). Public, like catalog covers.",
        "operationId": "getFileCover",
        "parameters": [
          {
            "in": "path",
            "name": "sha256",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Sha256",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/*": {}
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No such file, or no cover in it"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get File Cover",
        "tags": [
          "library"
        ]
      }
    },
...
    "/v1/library": {
      "get": {
        "description": "The books of the signed-in reader, most recent first.",
        "operationId": "getLibrary",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LibraryItemResponse"
                  },
                  "title": "Response Getlibrary",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Library",
        "tags": [
          "library"
        ]
      }
    },
    "/v1/library/files": {
      "post": {
        "description": "Import an EPUB, PDF, CBZ or CBR file into the library.",
        "operationId": "importFile",
        "parameters": [
          {
            "in": "header",
            "name": "X-Babel-Device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Babel-Device"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_importFile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Import File",
        "tags": [
          "library"
        ]
      }
    },
    "/v1/library/files/{sha256}": {
      "post": {
        "description": "Add a file already on Babel to the library, without uploading it again.",
        "operationId": "addStoredFile",
        "parameters": [
          {
            "in": "path",
            "name": "sha256",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Sha256",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Babel-Device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Babel-Device"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Stored File",
        "tags": [
          "library"
        ]
      }
    },
    "/v1/library/follows": {
      "get": {
        "description": "Unfinished AO3 works imported by link, checked daily for new chapters.",
        "operationId": "getFollows",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FollowResponse"
                  },
                  "title": "Response Getfollows",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Follows",
        "tags": [
          "library"
        ]
      }
    },
    "/v1/library/follows/{follow_id}": {
      "delete": {
        "description": "Stop checking this book for new chapters (the book stays).",
        "operationId": "stopFollow",
        "parameters": [
          {
            "in": "path",
            "name": "follow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Follow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stop Follow",
        "tags": [
          "library"
        ]
      }
    },
    "/v1/library/follows/{follow_id}/check": {
      "post": {
        "description": "Look for new chapters now; a new version replaces the book's file.",
        "operationId": "checkFollow",
        "parameters": [
          {
            "in": "path",
            "name": "follow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Follow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Check Follow",
        "tags": [
          "library"
        ]
      }
    },
...
    "/v1/library/links": {
      "post": {
        "description": "Import the book a link points to (AO3 works are fetched at AO3's pace).",
        "operationId": "importLink",
        "parameters": [
          {
            "in": "header",
            "name": "X-Babel-Device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Babel-Device"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Import Link",
        "tags": [
          "library"
        ]
      }
    },
    "/v1/library/links/preview": {
      "post": {
        "description": "What a pasted link points to: an AO3 work, a Gutenberg book or a file.",
        "operationId": "previewLink",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Preview Link",
        "tags": [
          "library"
        ]
      }
    },
    "/v1/library/paper": {
      "post": {
        "description": "A book you own on paper, to follow your reading without a file. If the work is\nalready (or was) in your library, that book is marked as owned on paper instead.",
        "operationId": "addPaperBook",
        "parameters": [
          {
            "in": "header",
            "name": "X-Babel-Device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Babel-Device"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperBookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Paper Book",
        "tags": [
          "library"
        ]
      }
    },
    "/v1/library/{item_id}": {
      "delete": {
        "description": "Take a book out of the library. Its status, review, notes and positions are kept and\ncome back if the same file is added again.",
        "operationId": "removeFromLibrary",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Babel-Device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Babel-Device"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove From Library",
        "tags": [
          "library"
        ]
      }
    },
...
    "/v1/library/{item_id}/positions": {
      "get": {
        "description": "Where each device stopped in this book, most recent first.",
        "operationId": "getReadingPositions",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReadingPositionResponse"
                  },
                  "title": "Response Getreadingpositions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Positions",
        "tags": [
          "library"
        ]
      }
    },
...
  }
}